// src/lib/api/model/emptyTimeSettingRequestDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Unplan API
 * Unplan Server API Documentation
 * OpenAPI spec version: v1.0.0
 */
import type { RecommendBanTime } from './recommendBanTime';

export interface EmptyTimeSettingRequestDto {
  is_empty_time_recommend_on?: boolean;
  empty_time_criteria_minutes?: number;
  is_recommend_ban_time_on?: boolean;
  recommend_ban_times?: RecommendBanTime[];
}